Let alt be a float for benefit of D108 and D109 protocols to make more immunune from rounding resulting from integer truncation. (Why do I sense 'm going to regret doing this so late in the release cycle?)

+++ trunk/gpsbabel/jeeps/gps.h
@@ -129,7 +129,7 @@ typedef struct GPS_SWay
     int32  colour;
     char   cc[2];
     UC     wpt_class;
-    int32  alt;
+    float  alt;
     char   city[24];
     char   state[2];
     char   name[30];
